[EI-152] [api] refactor: baseEntity 추가
createdAt, updatedAt column
