+++ api/src/numerical-guidance/infrastructure/adapter/persistence/base.entity.ts
@@ -1,0 +1,9 @@
+import { CreateDateColumn, UpdateDateColumn } from 'typeorm';
+
+export class BaseEntity {
+  @CreateDateColumn()
+  createdAt: Date;
+
+  @UpdateDateColumn()
+  updatedAt: Date;
+}
