+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-board-metadata/mapper/indicator-board-metadata.mapper.ts
@@ -1,6 +1,6 @@
 import { IndicatorBoardMetadataEntity } from '../entity/indicator-board-metadata.entity';
-import { IndicatorBoardMetadata } from '../../../../domain/indicator-board-metadata';
-import { MemberEntity } from '../../../../../auth/member.entity';
+import { IndicatorBoardMetadata } from '../../../../../domain/indicator-board-metadata';
+import { MemberEntity } from '../../../../../../auth/member.entity';
 
 export class IndicatorBoardMetadataMapper {
   static mapDomainToEntity(indicatorBoardMetaData: IndicatorBoardMetadata, member: MemberEntity) {
@@ -8,6 +8,8 @@ export class IndicatorBoardMetadataMapper {
     indicatorBoardMetaDataEntity.indicatorBoardMetaDataName = indicatorBoardMetaData.indicatorBoardMetaDataName;
     indicatorBoardMetaDataEntity.tickers = indicatorBoardMetaData.tickers;
     indicatorBoardMetaDataEntity.member = member;
+    indicatorBoardMetaDataEntity.createdAt = indicatorBoardMetaData.createdAt;
+    indicatorBoardMetaDataEntity.updatedAt = indicatorBoardMetaData.updatedAt;
     return indicatorBoardMetaDataEntity;
   }
