+++ api/src/numerical-guidance/infrastructure/adapter/persistence/indicator-list/entity/indicator.entity.ts
@@ -1,5 +1,6 @@
 import { IndicatorType, Market } from 'src/utils/type/type-definition';
-import { BaseEntity, Column, Entity, PrimaryColumn } from 'typeorm';
+import { Column, Entity, PrimaryColumn } from 'typeorm';
+import { BaseEntity } from '../../base.entity';
 
 @Entity()
 export class IndicatorEntity extends BaseEntity {
